+++ e2etest/container/container.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"regexp"
 	"strconv"
+	"strings"
 	"testing"
 	"time"
 
@@ -189,7 +190,14 @@ func (m *Manager) RunBabylondResource(
 	baseHeaderHex string,
 	slashingPkScript string,
 	epochInterval uint,
+	covenantPks ...*btcec.PublicKey,
 ) (*dockertest.Resource, error) {
+	covenantPksStr := make([]string, len(covenantPks))
+	for i, cvPk := range covenantPks {
+		covenantPksStr[i] = bbn.NewBIP340PubKeyFromBTCPK(cvPk).MarshalHex()
+	}
+	quorumSize := (len(covenantPks) / 2) + 1
+
 	cmd := []string{
 		"sh", "-c", fmt.Sprintf(
 			"babylond testnet --v=1 --output-dir=/home --starting-ip-address=192.168.10.2 "+
@@ -197,9 +205,9 @@ func (m *Manager) RunBabylondResource(
 				"--btc-confirmation-depth=3 --additional-sender-account --btc-network=regtest "+
 				"--min-staking-time-blocks=200 --min-staking-amount-sat=10000 "+
 				"--epoch-interval=%d --slashing-pk-script=%s --btc-base-header=%s "+
-				"--covenant-quorum=1 --covenant-pks=%s && chmod -R 777 /home && "+
+				"--covenant-quorum=%d --covenant-pks=%s --max-finality-providers-in-script=2 && chmod -R 777 /home && "+
 				"export BABYLON_BLS_PASSWORD=password && babylond start --home=/home/node0/babylond",
-			epochInterval, slashingPkScript, baseHeaderHex, bbn.NewBIP340PubKeyFromBTCPK(covenantPK).MarshalHex()),
+			epochInterval, slashingPkScript, baseHeaderHex, quorumSize, strings.Join(covenantPksStr, ",")),
 	}
 
 	resource, err := m.pool.RunWithOptions(